% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_friendship_generator.R
\name{percent_friendship_generator}
\alias{percent_friendship_generator}
\title{Percent friendship generator}
\usage{
percent_friendship_generator(name1, name2)
}
\arguments{
\item{name1}{name of first friend}

\item{name2}{name of second friend}
}
\value{
percent friendship level between the two people
}
\description{
Percent friendship generator
}
\examples{
percent_friendship_generator("erika", "ale")
}
