% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soulmate_name_start_with.R
\name{soulmate_name_starts_with}
\alias{soulmate_name_starts_with}
\title{What letter does soulmate name start with}
\usage{
soulmate_name_starts_with(your_name, fave_food)
}
\arguments{
\item{your_name}{your name}

\item{fave_food}{name of your favorite food}
}
\value{
letter your soulmate's name starts with
}
\description{
What letter does soulmate name start with
}
\examples{
soulmate_name_starts_with("erika", "sushi")
}
